refactor: Load Knockout locally
Loading Knockout locally instead of through cdn so that I can see
errors that come up when data can't be loaded. When loading
Knockout from a cdn and there is no network connection, it won't
do a zip code search at all because the search button is using
a data-bind.