feat: Add and remove map markers
- Added code to create a Google Map to View Model so it
would be in the same scope as the code to create Markers
for the map. This caused the objects hidden by Knockout
to flash on the screen when page loads, so had to create a
hidden class to hide them until Knockout is loaded.
- Markers are stored in the MarketList array because it
should be easier to work with them that way.