fix: Fix InfoWindow being hidden behind sidebar
InfoWindow was hidden behind sidebar on larger screens where
the sidebar isn't hidden when a market is selected. Made it
so it does not pan to the marker on smaller screens since the
sidebar hides and it looks jumpy to pan on a small screen.
Also set max-width for InfoWindow because it was too wide
on larger screens.